Initial import
This is a minimum bootstrap environment for bare-metal rust on
Teensy4. It contains the necessary static data structures and linker
script to load a simple program. There are not yet any hardware
bindings, so direct register access is needed in order to do anything.

This includes an example which uses direct register access to enable
the Teensy's orange LED.