Add debug helpers
This adds a simple debug helper using GPIO7 to drive pins 12 and 13.

Pin 13 is the built-in LED, and in the bootup example is used in the
panic handler. Pin 12 should be wired to another LED to indicate
whatever is needed during development.

In theory more pins could be supported. I've got a 10-segment bar
graph that I might try to use to indicate the boot progress.