Minimize usage of `debug`
This removes all usages of debug outside of the panic handlers (which
should never be hit in the examples as-written). This ensures that
debug mode doesn't interfere with other attempts to use the pins as
hardware is implemented. Unless something has gone *very* wrong, the
LpUart should now be used for debugging purposes.